Runing wollok in p5
